pyproject.toml
[tool.poetry]
name = "pythonproject-1-project"
version = "0.1.0"
description = ""
authors = ["alex <mmclinks.ml@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
mport = "^0.1.1"
logging = "^0.4.9.6"
flask = "^3.0.3"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.mypy_cache
    | \.venv
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
